ci: :fire: create task comments and user group all

+++ src/app/modules/user/user.service.ts
@@ -2,8 +2,9 @@ import { IUser } from './user.interface';
 import { User } from './user.model';
 import ApiError from '../../../utils/ApiError';
 import httpStatus from 'http-status';
+import { deleteFile } from '../../../utils/deleteFile';
 
-const createUser = async (payload: IUser): Promise<IUser> => {
+const createUser = async (payload: IUser): Promise<Omit<IUser, 'password'>> => {
   const user = new User();
   const isUserExist = await user.isUserExist(payload.email);
   if (isUserExist) {
@@ -12,8 +13,7 @@ const createUser = async (payload: IUser): Promise<IUser> => {
   const result = await User.create(payload);
   if (!result)
     throw new ApiError(httpStatus.BAD_REQUEST, 'Failed to create user');
-  const userWithoutPassword = result.toObject();
-  delete userWithoutPassword.password;
+  const { password, ...userWithoutPassword } = result.toObject();
   return userWithoutPassword;
 };
 
@@ -34,7 +34,9 @@ const updateUser = async (
 ): Promise<IUser | null> => {
   const findUser = await User.findById(id);
   if (!findUser) throw new ApiError(httpStatus.NOT_FOUND, 'User not found');
-
+  if (payload.profileImage && findUser.profileImage) {
+    deleteFile(findUser.profileImage);
+  }
   const result = await User.findOneAndUpdate({ _id: id }, payload, {
     new: true,
     runValidators: true,
@@ -44,6 +46,11 @@ const updateUser = async (
 };
 
 const deleteUser = async (id: string): Promise<IUser | null> => {
+  const findUser = await User.findById(id);
+  if (!findUser) throw new ApiError(httpStatus.NOT_FOUND, 'User not found');
+  if (findUser.profileImage) {
+    deleteFile(findUser.profileImage);
+  }
   const result = await User.findOneAndDelete({ _id: id });
   return result;
 };
